justto/requirements.txt
Django==1.11.15
django-hstore==1.4.2
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-hstore==1.3
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dotenv==0.9.1
pytz==2018.5
